requirements.txt
asgiref==3.11.0
attrs==25.4.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.12.0
django-redis==6.0.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
execnet==2.1.2
freezegun==1.5.5
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-mock==1.12.1
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
gunicorn==25.0.3
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
